reopened

+++ src/main/java/net/feusalamander/betterskills/procedures/AbilityconfigprocProcedure.java
@@ -57,7 +57,7 @@ public class AbilityconfigprocProcedure {
 		{
 			Entity _ent = entity;
 			if (_ent instanceof ServerPlayerEntity) {
-				BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+				BlockPos _bpos = new BlockPos(x, y, z);
 				NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 					@Override
 					public ITextComponent getDisplayName() {
